+++ Sources/BugsnagPerformance/Private/Tracer.mm
@@ -149,7 +149,7 @@ Tracer::startNetworkSpan(NSString *httpMethod, SpanOptions options) noexcept {
 BugsnagPerformanceSpan *
 Tracer::startViewLoadPhaseSpan(NSString *className,
                                NSString *phase,
-                               BugsnagPerformanceSpan *parentContext) noexcept {
+                               BugsnagPerformanceSpanContext *parentContext) noexcept {
     NSString *name = [NSString stringWithFormat:@"[ViewLoadPhase/%@]/%@", phase, className];
     SpanOptions options;
     options.parentContext = parentContext;