+++ Sources/BugsnagPerformance/include/BugsnagPerformance/BugsnagPerformanceSpan.h
@@ -5,28 +5,20 @@
 //  Created by Nick Dowell on 23/09/2022.
 //
 
-#import <Foundation/Foundation.h>
+#import <BugsnagPerformance/BugsnagPerformanceSpanContext.h>
 
 NS_ASSUME_NONNULL_BEGIN
 
-typedef union {
-    __uint128_t value;
-    struct {
-        uint64_t lo;
-        uint64_t hi;
-    };
-} TraceId;
-
-typedef uint64_t SpanId;
-
 OBJC_EXPORT
-@interface BugsnagPerformanceSpan : NSObject
+@interface BugsnagPerformanceSpan : BugsnagPerformanceSpanContext
 
-@property(nonatomic,readonly) TraceId traceId;
-@property(nonatomic,readonly) SpanId spanId;
 @property(nonatomic,readonly) BOOL isValid;
 
 - (instancetype)init NS_UNAVAILABLE;
+
+- (instancetype) initWithTraceId:(TraceId)traceId spanId:(SpanId)spanId NS_UNAVAILABLE;
+
+- (instancetype) initWithTraceIdHi:(uint64_t)traceIdHi traceIdLo:(uint64_t)traceIdLo spanId:(SpanId)spanId NS_UNAVAILABLE;
 
 - (void)abortIfOpen;