+++ features/fixtures/ios/Fixture/FixtureConfig.swift
@@ -9,6 +9,7 @@ import Foundation
 
 class FixtureConfig {
     let mazeRunnerURL: URL
+    let docsURL: URL
     let tracesURL: URL
     let commandURL: URL
     let metricsURL: URL
@@ -18,6 +19,7 @@ class FixtureConfig {
 
     init(mazeRunnerBaseAddress: URL) {
         mazeRunnerURL = mazeRunnerBaseAddress
+        docsURL = mazeRunnerBaseAddress.appendingPathComponent("docs")
         tracesURL = mazeRunnerBaseAddress.appendingPathComponent("traces")
         commandURL = mazeRunnerBaseAddress.appendingPathComponent("command")
         metricsURL = mazeRunnerBaseAddress.appendingPathComponent("metrics")