+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 Changelog
 =========
 
+## 1.6.2 (2024-07-04)
+
+### Bug fixes
+
+* Handle case where the user manually sets the network callback to nil. [281](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/281)
+
+### Enhancements
+
+* Span parentage APIs now take a BugsnagPerformanceSpanContext, which BugsnagPerformanceSpan is now a subclass of. You no longer need to assign a BugsnagPerformanceSpan as a parent. [280](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/280)
+
 ## 1.6.1 (2024-06-24)
 
 ### Bug fixes